+++ mobile/android/geckoview/src/main/java/org/mozilla/geckoview/WebExtensionController.java
@@ -247,29 +247,6 @@ public class WebExtensionController {
    */
   @UiThread
   public interface PromptDelegate {
-    /**
-     * @deprecated Please use onInstallPromptRequest instead. Called whenever a new extension is
-     *     being installed. This is intended as an opportunity for the app to prompt the user for
-     *     the permissions required by this extension.
-     * @param extension The {@link WebExtension} that is about to be installed. You can use {@link
-     *     WebExtension#metaData} to gather information about this extension when building the user
-     *     prompt dialog.
-     * @param permissions The list of permissions that are granted during installation.
-     * @param origins The list of origins that are granted during installation.
-     * @return A {@link GeckoResult} that completes to either {@link AllowOrDeny#ALLOW ALLOW} if
-     *     this extension should be installed or {@link AllowOrDeny#DENY DENY} if this extension
-     *     should not be installed. A null value will be interpreted as {@link AllowOrDeny#DENY
-     *     DENY}.
-     */
-    @Nullable
-    @Deprecated
-    @DeprecationSchedule(id = "web-extension-on-install-prompt", version = 134)
-    default GeckoResult<AllowOrDeny> onInstallPrompt(
-        @NonNull final WebExtension extension,
-        @NonNull final String[] permissions,
-        @NonNull final String[] origins) {
-      return null;
-    }
 
     /**
      * Called whenever a new extension is being installed. This is intended as an opportunity for
@@ -1199,38 +1176,20 @@ public class WebExtensionController {
       return;
     }
 
-    @SuppressWarnings("deprecation")
-    final GeckoResult<AllowOrDeny> promptResponseDeprecated =
-        mPromptDelegate.onInstallPrompt(
+    final GeckoResult<WebExtension.PermissionPromptResponse> promptResponse =
+        mPromptDelegate.onInstallPromptRequest(
             extension, message.getStringArray("permissions"), message.getStringArray("origins"));
-    // To be deleted after addressing https://bugzilla.mozilla.org/show_bug.cgi?id=1919374
-    // If we get null from onInstallPrompt this means nobody has implemented it, so we proceed to
-    // call the new API onInstallPromptRequest.
-    if (promptResponseDeprecated != null) {
-      callback.resolveTo(
-          promptResponseDeprecated.map(
-              allowOrDeny -> {
-                final GeckoBundle response = new GeckoBundle(2);
-                response.putBoolean("allow", AllowOrDeny.ALLOW.equals(allowOrDeny));
-                response.putBoolean("privateBrowsingAllowed", false);
-                return response;
-              }));
-    } else {
-      final GeckoResult<WebExtension.PermissionPromptResponse> promptResponse =
-          mPromptDelegate.onInstallPromptRequest(
-              extension, message.getStringArray("permissions"), message.getStringArray("origins"));
-      if (promptResponse == null) {
-        return;
-      }
-      callback.resolveTo(
-          promptResponse.map(
-              userResponse -> {
-                final GeckoBundle response = new GeckoBundle(2);
-                response.putBoolean("allow", userResponse.isPermissionsGranted);
-                response.putBoolean("privateBrowsingAllowed", userResponse.isPrivateModeGranted);
-                return response;
-              }));
+    if (promptResponse == null) {
+      return;
     }
+    callback.resolveTo(
+        promptResponse.map(
+            userResponse -> {
+              final GeckoBundle response = new GeckoBundle(2);
+              response.putBoolean("allow", userResponse.isPermissionsGranted);
+              response.putBoolean("privateBrowsingAllowed", userResponse.isPrivateModeGranted);
+              return response;
+            }));
   }
 
   private void updatePrompt(final GeckoBundle message, final EventCallback callback) {